+++ example/src/custom/BackgroundAudioSubMenu.tsx
@@ -1,6 +1,8 @@
-import React, { useContext, useState } from 'react';
+import { CustomSubMenu, Option } from './CustomSubMenu';
+import * as React from 'react';
 import type { StyleProp, ViewStyle } from 'react-native';
-import { MenuRadioButton, MenuView, PlayerContext, ScrollableMenu, SubMenuWithButton } from '@theoplayer/react-native-ui';
+import { useContext } from 'react';
+import { PlayerContext } from '@theoplayer/react-native-ui';
 
 export interface BackgroundAudioSubMenuProps {
   /**
@@ -9,61 +11,19 @@ export interface BackgroundAudioSubMenuProps {
   menuStyle?: StyleProp<ViewStyle>;
 }
 
-interface Option {
-  label: string;
-  value: boolean;
-}
-
-const OptionDisabled = { label: 'Disabled', value: false };
-const OptionEnabled = { label: 'Enabled', value: true };
-const options: Option[] = [OptionDisabled, OptionEnabled];
-
 /**
  * A button component that opens a backgroundAudio selection menu for the `react-native-theoplayer` UI.
  */
-export const BackgroundAudioSubMenu = (props: BackgroundAudioSubMenuProps) => {
-  const { menuStyle } = props;
-  const createMenu = () => {
-    return <BackgroundAudioSelectionView style={menuStyle} />;
-  };
-  const player = useContext(PlayerContext).player;
-  const selectedLabel = player.backgroundAudioConfiguration.enabled ? OptionEnabled.label : OptionDisabled.label;
-  return <SubMenuWithButton menuConstructor={createMenu} label={'Bg Audio'} preview={selectedLabel} />;
-};
+export const BackgroundAudioSubMenu = (props?: BackgroundAudioSubMenuProps) => {
+  const ctx = useContext(PlayerContext);
 
-export interface BackgroundAudioSelectionViewProps {
-  style?: StyleProp<ViewStyle>;
+  return <CustomSubMenu
+    title={'Background Audio'}
+    menuStyle={props?.menuStyle}
+    label={'Bg Audio'}
+    options={[{ label: 'Disabled', value: false }, { label: 'Enabled', value: true }]}
+    onOptionSelected={(option: Option<boolean>) => {
+      ctx.player.backgroundAudioConfiguration = { enabled: option.value };
+    }}
+    currentOption={() => ctx.player.backgroundAudioConfiguration.enabled ?? false} />;
 }
-
-const BackgroundAudioSelectionView = (props: BackgroundAudioSelectionViewProps) => {
-  const { style } = props;
-  const player = useContext(PlayerContext).player;
-
-  const [selected, setSelected] = useState<Option>(player.backgroundAudioConfiguration.enabled ? OptionEnabled : OptionDisabled);
-
-  const onSelectOption = (id: number | undefined): void => {
-    if (id !== undefined) {
-      player.backgroundAudioConfiguration = { enabled: options[id].value };
-      setSelected(options[id]);
-    }
-  };
-
-  return (
-    <MenuView
-      style={style}
-      menu={
-        <ScrollableMenu
-          title={'Background Audio'}
-          items={options.map((option, id) => (
-            <MenuRadioButton
-              key={id}
-              label={option.label}
-              uid={id}
-              onSelect={onSelectOption}
-              selected={selected.value === option.value}></MenuRadioButton>
-          ))}
-        />
-      }
-    />
-  );
-};